Impl OBJ model loading w/assimp
Added support for assimp from lwjgl. Impl loadModel() in the new class ModelLoader -- loads the given OBJ file and creates a custom mesh (class: MeshFromOBJ) using the data in the file. Currently only works with files which have vertex positions, normals and faces for sure. Also accepts those with texture coordinates. Used it to draw a dragon on screen!! :)