Cont w/reflective materials
Created the class ReflectiveMaterial, which represents a material withe full or partial reflection. If a texture list is used, then a reflection map must be one of the textures. If plain colours are used instead, there is the option to specify a 'reflectivity coeff' (K_refl) to determine how much of the skybox it would reflect (e.g. as seen in the floor atm). (note: I made this 'reflectivity coefficient' name up, just to be able to vary the amount of reflection from an object). Note: also added the 'isReflectiveMaterial' uniform to shader to render material appropriately.