Created shaders that impl Blinn-Phong shading
Created shaders that use the Blinn-Phong lighting model, rather than the Phong model. (also changed a bit the params to the calc lights functions in the latest phong shader: e.g. diffColour -> diffComponent (which is diffColour * material.K_diff) -- done to avoid unnecessary duplication of this calc in each funct)